Added the function to retrieve individual documents.
It includes also a set of functions to transform between Map-WritableMap
and Array-WritableArray. This is because CouchBase is quite #yolo
(everything is a Map<String, Object>) and React Native is very picky (it
only accepts native types and they need to be carefully added one by one
depending on the type).